finalização das implementações

+++ API/StudentGroupsManager/Controllers/GroupController.cs
@@ -11,11 +11,17 @@ public class GroupController : ControllerBase
     private readonly ILogger<GroupController> _logger;
     private readonly ICourseGroupRepository _courseGroupRepository;
     private readonly IStudentGroupRepository _studentGroupRepository;
+    private readonly IParametrosRepository _parametrosRepository;
 
-    public GroupController(ICourseGroupRepository courseGroupRepository, IStudentGroupRepository studentGroupRepository, ILogger<GroupController> logger)
+    public GroupController(
+        ICourseGroupRepository courseGroupRepository, 
+        IStudentGroupRepository studentGroupRepository, 
+        IParametrosRepository parametrosRepository,
+        ILogger<GroupController> logger)
     {
         _courseGroupRepository = courseGroupRepository;
         _studentGroupRepository = studentGroupRepository;
+        _parametrosRepository = parametrosRepository;
         _logger = logger;
     }
 
@@ -29,7 +35,7 @@ public class GroupController : ControllerBase
     /// GET: api/group
     [HttpGet()]
     public ActionResult<IEnumerable<GroupGetDTO>> GetAll()
-        => Ok(_courseGroupRepository.GetAll());
+        => Ok(_courseGroupRepository.GetAllWithDto());
     
     /// <summary>
     /// Obtém lista de grupos ativos por grupo
@@ -55,6 +61,12 @@ public class GroupController : ControllerBase
     [HttpPost]
     public ActionResult CreateGroup(GroupCreateDTO dto)
     {
+        if (_parametrosRepository.DeadLineReachedByCourse(dto.CourseId))
+        {
+            _logger.LogInformation("Não é possível criar um novo grupo pois a data limite foi atingida");
+            return BadRequest("Data limite para criação do grupo atingida");
+        }
+        
         _courseGroupRepository.CreateAGroup(dto);
         return Ok();
     }
@@ -62,17 +74,23 @@ public class GroupController : ControllerBase
     /// <summary>
     /// Altera o número máximo de estudantes por grupo
     /// </summary>
-    /// <param name="courseId"></param>
+    /// <param name="groupId"></param>
     /// <param name="numberOfStudents"></param>
     /// <returns></returns>
     /// <response code="200">Retorna Sucesso</response>
     /// <response code="401">Não Autenticado</response>
     /// <response code="403">Não Autorizado</response>
     /// PUT: api/group/1/3
-    [HttpPut("{courseId:int}/{numberOfStudents:int}")]
-    public ActionResult ChangeGroupMembersLimit(int courseId, int numberOfStudents)
+    [HttpPut("{groupId:int}/{numberOfStudents:int}")]
+    public ActionResult ChangeGroupMembersLimit(int groupId, int numberOfStudents)
     {
-        _courseGroupRepository.ChangeMaxNumberOfStudents(courseId, numberOfStudents);
+        if (_parametrosRepository.DeadLineReachedByGroup(groupId))
+        {
+            _logger.LogInformation("Não é possível alterar o quantidade de membros pois a data limite foi atingida");
+            return BadRequest("Data limite para alterar a quantiodade de membros atingida");
+        };
+        
+        _courseGroupRepository.ChangeMaxNumberOfStudents(groupId, numberOfStudents);
         return Ok();
     }  
     
@@ -89,6 +107,12 @@ public class GroupController : ControllerBase
     [HttpPost("{groupId:int}/{studentId:int}")]
     public ActionResult EnrollAGroup(int groupId, int studentId)
     {
+        if (_parametrosRepository.DeadLineReachedByGroup(groupId))
+        {
+            _logger.LogInformation("Não é possível entrar em um grupo pois a data limite foi atingida");
+            return BadRequest("Data limite para ingressar no grupo atingida");
+        };
+        
         try
         {
             _courseGroupRepository.EnrollAGroup(groupId);
@@ -115,8 +139,23 @@ public class GroupController : ControllerBase
     [HttpDelete("{groupId:int}/{studentId:int}")]
     public ActionResult UnEnrollAGroup(int groupId, int studentId)
     {
-        _studentGroupRepository.UnEnrollStudent(groupId, studentId);
-        return Ok();
+        if (_parametrosRepository.DeadLineReachedByGroup(groupId))
+        {
+            _logger.LogInformation("Não é possível sair de um grupo pois a data limite foi atingida");
+            return BadRequest("Data limite para sair do grupo atingida");
+        };
+        
+        try
+        {
+            _courseGroupRepository.UnEnrollAGroup(groupId, studentId);
+            _studentGroupRepository.UnEnrollStudent(groupId, studentId);
+            return Ok();
+        }
+        catch (Exception)
+        {
+            _logger.LogError("Ocorreu um erro ao deixar o grupo");
+            return BadRequest();
+        }
     }
     
     /// <summary>
@@ -131,6 +170,12 @@ public class GroupController : ControllerBase
     [HttpDelete("{groupId:int}")]
     public async Task<ActionResult> Delete(int groupId)
     {
+        if (_parametrosRepository.DeadLineReachedByGroup(groupId))
+        {
+            _logger.LogInformation("Não é possível excluir o grupo pois a data limite foi atingida");
+            return BadRequest("Data limite exclusão do grupo atingida");
+        };
+        
         await _studentGroupRepository.UnEnrollAllStudents(groupId);
         var group = _courseGroupRepository.GetById(groupId);
         _courseGroupRepository.Delete(group);
